# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "OKTIOAPI",
    "description": "OKTIO API for reverse auctions",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "x-auth-token",
      "scheme": "bearer",
      "in": "header"
    }
  },
...
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/api/seller/getAllMyAuctions/{email}": {
      "get": {
        "tags": ["Seller"],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "Fetch all auction for a specific seller",
            "schema": {
              "$ref": "#/definitions/getAllMyAuctions"
            }
          }
        ],
        "summary": "Fetch all auction for a specific seller",
        "responses": {
          "200": {
            "description": "Successful"
          }
        }
      }
    },
...
    "/api/seller/supplier/invitation/{id}": {
      "post": {
        "tags": ["Seller"],
        "summary": "Accpet Company Supplier Invitation",
        "parameters": [
          {
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfull"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/seller/getCompanyInvitation": {
      "post": {
        "tags": ["Seller"],
        "summary": "Get all invitation from company",
        "parameters": [
          {
            "name": "seller",
            "in": "body",
            "description": "Get all company invitation",
            "schema": {
              "$ref": "#/definitions/getCompanyInvitation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/getCompanyInvitation"
            }
          }
        }
      }
    },
    "/api/seller/getCompanyAcceptedInvitation": {
      "post": {
        "tags": ["Seller"],
        "summary": "Get only Accepted invitation from company",
        "parameters": [
          {
            "name": "seller",
            "in": "body",
            "description": "Get only Accepted invitation from company",
            "schema": {
              "$ref": "#/definitions/getCompanyAcceptedInvitation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/getCompanyAcceptedInvitation"
            }
          }
        }
      }
    },
    "/api/seller/acceptorDeclineCompanyInvitation": {
      "patch": {
        "tags": ["Seller"],
        "summary": "Accept or Decline invitation from company",
        "parameters": [
          {
            "name": "seller",
            "in": "body",
            "description": "Accept or decline invitation",
            "schema": {
              "$ref": "#/definitions/acceptorDeclineCompanyInvitation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/acceptorDeclineCompanyInvitation"
            }
          }
        }
      }
    },
    "/api/seller/auctionInvitation": {
      "post": {
        "tags": ["Seller"],
        "summary": "Register and send invitation email to a seller by a company in the system",
        "parameters": [
          {
            "name": "seller",
            "in": "body",
            "description": "Seller that we want to create",
            "schema": {
              "$ref": "#/definitions/Seller"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/Seller"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "Seller": {
      "required": ["auctionId", "buyerUserId", "email", "company_buyer_name"],
      "properties": {
        "auctionId": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "buyerUserId": {
          "type": "string"
        },
        "company_buyer_name": {
          "type": "string"
        }
      }
    },
    "getCompanyInvitation": {
      "required": ["email"],
      "properties": {
        "email": {
          "type": "string"
        }
      }
    },
    "getCompanyAcceptedInvitation": {
      "required": ["email"],
      "properties": {
        "email": {
          "type": "string"
        }
      }
    },
    "acceptorDeclineCompanyInvitation": {
      "required": ["id", "invitation"],
      "properties": {
        "id": {
          "type": "string"
        },
        "invitation": {
          "type": "string"
        }
      }
    },
...
    "getAllMyAuctions": {
      "required": ["sellersEmail"],
      "properties": {
        "sellersEmail": {
          "type": "string"
        }
      }
    },
...
  }
}